Adicionar serializers indiividuais no módulo serializer Remover serializer.py
