+++ api/serializer/sale.py
@@ -1,0 +1,9 @@
+from rest_framework import serializers
+from api.models.sale import Sale
+
+
+class SaleSerializer(serializers.ModelSerializer):
+    class Meta:
+        model = Sale
+
+        fields = ["pk", "seller", "year", "month", "amount", "comission_value"]
