+++ api/serializer/user.py
@@ -1,0 +1,9 @@
+from django.contrib.auth.models import User
+from rest_framework import serializers
+
+
+class UserSerializer(serializers.HyperlinkedModelSerializer):
+    class Meta:
+        model = User
+
+        fields = ["url", "username", "email", "is_staff"]
